# server/docs/swagger.yaml
definitions:
  models.Category:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/sql.NullTime'
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  sql.NullTime:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  structsv1.LinkCategoryRequest:
    properties:
      category_id:
        type: string
    required:
    - category_id
    type: object
  structsv1.ListCategoriesResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  structsv1.ListProductsResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      products:
        items:
          $ref: '#/definitions/structsv1.ResponseProduct'
        type: array
      total:
        type: integer
    type: object
  structsv1.MatchCategoriesResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
    type: object
  structsv1.RequestCategory:
    properties:
      name:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - name
    type: object
  structsv1.RequestProduct:
    properties:
      description:
        maxLength: 150
        type: string
      name:
        maxLength: 50
        minLength: 1
        type: string
      price:
        maximum: 999999999
        type: integer
      stock_quantity:
        maximum: 2147483647
        minimum: 1
        type: integer
    required:
    - description
    - name
    - price
    type: object
  structsv1.ResponseProduct:
    properties:
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/sql.NullTime'
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: integer
      stock_quantity:
        type: integer
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /categories:
    get:
      description: List categories with pagination, alphabetical by name
      parameters:
      - default: 20
        description: Page size (20|50|100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structsv1.ListCategoriesResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new category. This is the only way categories are created.
      parameters:
      - description: Category payload
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/structsv1.RequestCategory'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a category
      tags:
      - categories
  /categories/match:
    get:
      description: List active categories whose name starts with the given prefix
      parameters:
      - default: sale
        description: Name prefix
        in: query
        name: name
        type: string
      - default: 10
        description: Max results (5|10|20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structsv1.MatchCategoriesResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Match categories by name prefix
      tags:
      - categories
  /products:
    get:
      description: List products with pagination, newest first
      parameters:
      - default: 20
        description: Page size (20|50|100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structsv1.ListProductsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product. Categories are linked separately via the
        link route.
      parameters:
      - description: Product payload
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/structsv1.RequestProduct'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/structsv1.ResponseProduct'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a product
      tags:
      - products
  /products/{id}:
    delete:
      description: Soft-delete a product and its category relationships
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a product
      tags:
      - products
    get:
      description: Get a product by its UUID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structsv1.ResponseProduct'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update a product's own fields. Categories are managed via link/unlink
        routes.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product payload
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/structsv1.RequestProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structsv1.ResponseProduct'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a product
      tags:
      - products
  /products/{id}/categories/link:
    post:
      consumes:
      - application/json
      description: Link an existing category to a product. One category per call.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Category to link
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/structsv1.LinkCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structsv1.ResponseProduct'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Link a category to a product
      tags:
      - products
  /products/{id}/categories/unlink:
    post:
      consumes:
      - application/json
      description: Remove a category link from a product. One category per call.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Category to unlink
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/structsv1.LinkCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structsv1.ResponseProduct'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Unlink a category from a product
      tags:
      - products
  /products/name/{name}:
    get:
      description: Get a product by its name
      parameters:
      - description: Product name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structsv1.ResponseProduct'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a product by name
      tags:
      - products
swagger: "2.0"