# server/src/controllers/v1/ProductController.go
package controllersv1

import (
	"errors"

	"github.com/JayPonda/Product-catalog/server/src/services"
	v1 "github.com/JayPonda/Product-catalog/server/src/structs/v1"
	"github.com/JayPonda/Product-catalog/server/utils"
	"github.com/go-playground/validator/v10"
	"github.com/gofiber/fiber/v3"
	"github.com/google/uuid"
)

type ProductController struct {
	Service   *services.ProductService
	Validator *validator.Validate
}

func NewProductController(service *services.ProductService) *ProductController {
	return &ProductController{
		Service:   service,
		Validator: utils.NewValidator(),
	}
}

// ListProducts godoc
// @Summary      List products
// @Description  List products with pagination, newest first
// @Tags         products
// @Produce      json
// @Param        limit   query  int  false  "Page size (20|50|100)"  default(20)
// @Param        offset  query  int  false  "Offset"             default(0)
// @Success      200     {object}  v1.ListProductsResponse
// @Failure      400     {object}  map[string]string
// @Failure      500     {object}  map[string]string
// @Router       /products [get]
func (pc *ProductController) ListProducts(ctx fiber.Ctx) error {
	var query v1.ListProductsQuery

	if err := ctx.Bind().Query(&query); err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error":   "invalid query parameters",
			"details": err.Error(),
		})
	}

	if err := pc.Validator.Struct(query); err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error":   "validation failed",
			"details": err.Error(),
		})
	}

	if query.Limit == 0 {
		query.Limit = 20
	}

	response, err := pc.Service.ListProducts(query.Limit, query.Offset)
	if err != nil {
		return ctx.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": err.Error(),
		})
	}

	return ctx.JSON(response)
}

// CreateProduct godoc
// @Summary      Create a product
// @Description  Create a new product. Categories are linked separately via the link route.
// @Tags         products
// @Accept       json
// @Produce      json
// @Param        product  body      v1.RequestProduct  true  "Product payload"
// @Success      201      {object}  v1.ResponseProduct
// @Failure      400      {object}  map[string]string
// @Failure      409      {object}  map[string]string
// @Failure      500      {object}  map[string]string
// @Router       /products [post]
func (pc *ProductController) CreateProduct(ctx fiber.Ctx) error {
	var req v1.RequestProduct

	if err := ctx.Bind().JSON(&req); err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error":   "invalid request body",
			"details": err.Error(),
		})
	}

	if err := pc.Validator.Struct(req); err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error":   "validation failed",
			"details": err.Error(),
		})
	}

	product, err := pc.Service.CreateProduct(req)
	if err != nil {
		if errors.Is(err, services.ErrDuplicateProductName) {
			return ctx.Status(fiber.StatusConflict).JSON(fiber.Map{
				"error": err.Error(),
			})
		}
		return ctx.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": err.Error(),
		})
	}

	return ctx.Status(fiber.StatusCreated).JSON(product)
}

// GetProductById godoc
// @Summary      Get a product by ID
// @Description  Get a product by its UUID
// @Tags         products
// @Produce      json
// @Param        id   path      string  true  "Product ID"
// @Success      200  {object}  v1.ResponseProduct
// @Failure      400  {object}  map[string]string
// @Failure      404  {object}  map[string]string
// @Router       /products/{id} [get]
func (pc *ProductController) GetProductById(ctx fiber.Ctx) error {
	id, err := uuid.Parse(ctx.Params("id"))
	if err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": "invalid product id",
		})
	}

	product, err := pc.Service.GetProductById(id)
	if err != nil {
		return ctx.Status(fiber.StatusNotFound).JSON(fiber.Map{
			"error": "product not found",
		})
	}

	return ctx.JSON(product)
}

// GetProductByName godoc
// @Summary      Get a product by name
// @Description  Get a product by its name
// @Tags         products
// @Produce      json
// @Param        name   path      string  true  "Product name"
// @Success      200    {object}  v1.ResponseProduct
// @Failure      404    {object}  map[string]string
// @Router       /products/name/{name} [get]
func (pc *ProductController) GetProductByName(ctx fiber.Ctx) error {
	name := ctx.Params("name")

	product, err := pc.Service.GetProductByName(name)
	if err != nil {
		return ctx.Status(fiber.StatusNotFound).JSON(fiber.Map{
			"error": "product not found",
		})
	}

	return ctx.JSON(product)
}

// UpdateProduct godoc
// @Summary      Update a product
// @Description  Update a product's own fields. Categories are managed via link/unlink routes.
// @Tags         products
// @Accept       json
// @Produce      json
// @Param        id       path      string              true  "Product ID"
// @Param        product  body      v1.RequestProduct   true  "Product payload"
// @Success      200      {object}  v1.ResponseProduct
// @Failure      400      {object}  map[string]string
// @Failure      404      {object}  map[string]string
// @Failure      409      {object}  map[string]string
// @Failure      500      {object}  map[string]string
// @Router       /products/{id} [put]
func (pc *ProductController) UpdateProduct(ctx fiber.Ctx) error {
	id, err := uuid.Parse(ctx.Params("id"))
	if err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": "invalid product id",
		})
	}

	var req v1.RequestProduct

	if err := ctx.Bind().JSON(&req); err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error":   "invalid request body",
			"details": err.Error(),
		})
	}

	if err := pc.Validator.Struct(req); err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error":   "validation failed",
			"details": err.Error(),
		})
	}

	product, err := pc.Service.UpdateProduct(id, req)
	if err != nil {
		if errors.Is(err, services.ErrDuplicateProductName) {
			return ctx.Status(fiber.StatusConflict).JSON(fiber.Map{
				"error": err.Error(),
			})
		}
		return ctx.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": err.Error(),
		})
	}

	return ctx.JSON(product)
}

// LinkCategory godoc
// @Summary      Link a category to a product
// @Description  Link an existing category to a product. One category per call.
// @Tags         products
// @Accept       json
// @Produce      json
// @Param        id       path      string                    true  "Product ID"
// @Param        payload  body      v1.LinkCategoryRequest    true  "Category to link"
// @Success      200      {object}  v1.ResponseProduct
// @Failure      400      {object}  map[string]string
// @Failure      404      {object}  map[string]string
// @Failure      500      {object}  map[string]string
// @Router       /products/{id}/categories/link [post]
func (pc *ProductController) LinkCategory(ctx fiber.Ctx) error {
	id, err := uuid.Parse(ctx.Params("id"))
	if err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": "invalid product id",
		})
	}

	var req v1.LinkCategoryRequest

	if err := ctx.Bind().JSON(&req); err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error":   "invalid request body",
			"details": err.Error(),
		})
	}

	if err := pc.Validator.Struct(req); err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error":   "validation failed",
			"details": err.Error(),
		})
	}

	product, err := pc.Service.LinkCategory(id, req.CategoryID)
	if err != nil {
		if errors.Is(err, services.ErrProductNotFound) || errors.Is(err, services.ErrCategoryNotFound) {
			return ctx.Status(fiber.StatusNotFound).JSON(fiber.Map{
				"error": err.Error(),
			})
		}
		return ctx.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": err.Error(),
		})
	}

	return ctx.JSON(product)
}

// UnlinkCategory godoc
// @Summary      Unlink a category from a product
// @Description  Remove a category link from a product. One category per call.
// @Tags         products
// @Accept       json
// @Produce      json
// @Param        id       path      string                    true  "Product ID"
// @Param        payload  body      v1.LinkCategoryRequest    true  "Category to unlink"
// @Success      200      {object}  v1.ResponseProduct
// @Failure      400      {object}  map[string]string
// @Failure      404      {object}  map[string]string
// @Failure      500      {object}  map[string]string
// @Router       /products/{id}/categories/unlink [post]
func (pc *ProductController) UnlinkCategory(ctx fiber.Ctx) error {
	id, err := uuid.Parse(ctx.Params("id"))
	if err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": "invalid product id",
		})
	}

	var req v1.LinkCategoryRequest

	if err := ctx.Bind().JSON(&req); err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error":   "invalid request body",
			"details": err.Error(),
		})
	}

	if err := pc.Validator.Struct(req); err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error":   "validation failed",
			"details": err.Error(),
		})
	}

	product, err := pc.Service.UnlinkCategory(id, req.CategoryID)
	if err != nil {
		if errors.Is(err, services.ErrProductNotFound) || errors.Is(err, services.ErrCategoryNotFound) {
			return ctx.Status(fiber.StatusNotFound).JSON(fiber.Map{
				"error": err.Error(),
			})
		}
		return ctx.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": err.Error(),
		})
	}

	return ctx.JSON(product)
}

// DeleteProduct godoc
// @Summary      Delete a product
// @Description  Soft-delete a product and its category relationships
// @Tags         products
// @Param        id   path      string  true  "Product ID"
// @Success      204
// @Failure      400  {object}  map[string]string
// @Failure      500  {object}  map[string]string
// @Router       /products/{id} [delete]
func (pc *ProductController) DeleteProduct(ctx fiber.Ctx) error {
	id, err := uuid.Parse(ctx.Params("id"))
	if err != nil {
		return ctx.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": "invalid product id",
		})
	}

	err = pc.Service.DeleteProduct(id)
	if err != nil {
		return ctx.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": err.Error(),
		})
	}

	return ctx.Status(fiber.StatusNoContent).Send(nil)
}